fix(billling e profile): gerando os tickets e referenciando os repositórios, endpoints funcionando

+++ CentralTicket/Contexts/Auth/Entities/Sale.cs
@@ -5,7 +5,7 @@ namespace CentralTicket.Contexts.Auth.Entities
 {
     public class Sale : Base
     {
-        public double TotalValue { get; private set; }
+        public decimal TotalValue { get; private set; }
         public string PaymentMethod { get; set; }
         public string Status { get; set; }
         public string OrderCode { get; set; }
